requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.0
cryptography==3.0
decorator==4.4.2
Django==3.0.8
django-cors-headers==3.4.0
django-telegrambot==1.0.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
future==0.18.2
django-jalali-date==0.3.1
mysqlclient==2.0.1
Pillow==7.2.0
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.14.0
python-telegram-bot==12.8
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
tornado==6.0.4
